Update README.md
Add addToResponse into docs. 
There is a problem where in case that the service which header is forwarded to responds with another header with same name same value this plugin will duplicate the value on the response to the user. 

The fix is to disable the returning from traefik plugin level and let the downstream service take care of that